Before final submission

+++ src/main/java/ie/cit/video/JdbcVideoRepository.java
@@ -6,9 +6,13 @@ import java.util.List;
 
 import javax.sql.DataSource;
 
+import org.h2.jdbc.JdbcResultSet;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
+import org.springframework.security.access.annotation.Secured;
+import org.springframework.security.core.context.SecurityContextHolder;
 
+@Secured("ROLE_USER")
 public class JdbcVideoRepository {
 
 	private JdbcTemplate jdbcTemplate;
@@ -24,28 +28,34 @@ public class JdbcVideoRepository {
 	}
 
 	public void save(Videostore videostore) {
-		jdbcTemplate.update("insert into VIDEO (text, stocknum) values(?,?)",
-				videostore.getText(), videostore.getStocknum());
+
+		jdbcTemplate.update(
+				"insert into VIDEO (text, stocknum, owner) values(?,?,?)",
+				videostore.getText(), videostore.getStocknum(), getCurrentUser());
 	}
 
 	public Videostore get(int id) {
-		// System.out.println("cccc " +id);
+
 		return jdbcTemplate.queryForObject(
-				"select id, text, stocknum from VIDEO where id=?", new VideoStoreMapper(), id);
+				"select id, text, stocknum from VIDEO where id=? and owner=?", new VideoStoreMapper(), id, getCurrentUser());
 	}
 	
 	public List<Videostore> getAll(){
 		return jdbcTemplate.query(
-				"select id, text, stocknum from VIDEO",  new VideoStoreMapper());
+				"select id, text, stocknum from VIDEO where owner=?",  new VideoStoreMapper(), getCurrentUser());
 	}
 	
 	// DELETE, LOOK
 	public void delete(int id){
-		jdbcTemplate.update("delete from VIDEO where id=?", id);
+		jdbcTemplate.update("delete from VIDEO where id=? and owner=?", id, getCurrentUser());
 	}
 	
 	public void update(Videostore videostore){
-		jdbcTemplate.update("update VIDEO set text=?, stocknum=? where id=?", videostore.getText(), videostore.getStocknum(), videostore.getId());
+		jdbcTemplate.update("update VIDEO set text=?, stocknum=? where id=? and owner=?", videostore.getText(), videostore.getStocknum(), videostore.getId(), getCurrentUser());
+	}
+	
+	public String getCurrentUser(){
+		return SecurityContextHolder.getContext().getAuthentication().getName();
 	}
 }
